[Build tool] Added new target media to create tape program to be executed under BASIC environment [Crt0] Fixed size in the binary header for BIN_DISK and BIN_TAPE programs [lVGM] Fixed compilation warning [NEO mapper] Improved sample program

+++ projects/samples/s_lvgm.c
@@ -303,7 +303,7 @@ void ButtonLoop()
 void SetCursor(u8 id)
 {
 	g_CurrentButton = id % 6;
-	VDP_SetSpriteSM1(0, 8 + 16*g_CurrentButton, 128+8-1, g_ButtonEntry[g_CurrentButton].Char, COLOR_LIGHT_RED);
+	VDP_SetSpriteSM1(0, 8 + 16*g_CurrentButton, (u8)(128 + 8 - 1), g_ButtonEntry[g_CurrentButton].Char, COLOR_LIGHT_RED);
 }
 
 //-----------------------------------------------------------------------------
